requirements.txt
scrapy
scrapy-splash
PyYAML
requests
beautifulsoup4